Added memory detection, heap, memory information

+++ stdint.h
@@ -6,14 +6,14 @@
 #include "stddef.h"
 
 
-typedef unsigned    char    uint8_t;
-typedef             char    int8_t;
-typedef unsigned    short   uint16_t;
-typedef             short   int16_t;
-typedef unsigned    int     uint32_t;
-typedef             int     int32_t;
-typedef unsigned    long    uint64_t;
-typedef             long    int64_t;
+typedef unsigned    char        uint8_t;
+typedef             char        int8_t;
+typedef unsigned    short       uint16_t;
+typedef             short       int16_t;
+typedef unsigned    int         uint32_t;
+typedef             int         int32_t;
+typedef unsigned    long long   uint64_t;
+typedef             long long   int64_t;
 
 
 #endif
